fix: 优化上下文组件
