+++ app/renderer/src/main/src/pages/ai-agent/aiChatContent/AIContextToken/AIContextToken.tsx
@@ -1,44 +1,10 @@
-import { useCreation, useInViewport, useMemoizedFn } from 'ahooks'
-import {
-  AICostDetailsEcharts,
-  AICostDetailsEchartsProps,
-  AIPressureDetailsEcharts,
-  AIPressureDetailsEchartsProps,
-  ContextPressureEcharts,
-  ContextPressureEchartsProps,
-  ResponseSpeedEcharts,
-  ResponseSpeedEchartsProps,
-  TokenCountEcharts,
-  type ContextStatsChartMetric,
-} from '../../chatTemplate/AIEcharts'
+import React, { FC, memo } from 'react'
+import ContextPressurePanel from './ContextPressurePanel'
+import ContextCostPanel from './ContextCostPanel'
+import ContextTokenSummary from './ContextTokenSummary'
+import ContextDetailPopover from './ContextDetailPopover'
+import { YakitButtonProp } from '@/components/yakitUI/YakitButton/YakitButton'
 import styles from '../AIChatContent.module.scss'
-import { FC, memo, useCallback, useEffect, useRef, useState } from 'react'
-import { aiChatDataStore } from '../../store/ChatDataStore'
-import useChatIPCDispatcher from '@/pages/ai-agent/useContext/ChatIPCContent/useDispatcher'
-import { formatNumberUnits } from '../../utils'
-import {
-  OutlineArrowdownIcon,
-  OutlineArrowupIcon,
-  OutlinePresentationchartlineIcon,
-  OutlineXIcon,
-} from '@/assets/icon/outline'
-import classNames from 'classnames'
-import { useRafPolling } from '@/hook/useRafPolling/useRafPolling'
-import { cloneDeep, isEmpty } from 'lodash'
-import { getPressuresData, getCostData, getThreshold, getContextStatsData, isPerfDataChanged } from './utils'
-import { YakitButton, YakitButtonProp } from '@/components/yakitUI/YakitButton/YakitButton'
-import { YakitPopover } from '@/components/yakitUI/YakitPopover/YakitPopover'
-import { AIModelConfig } from '../../aiModelList/utils'
-import { getIconByAI } from '../../aiModelList/aiModelSelect/AIModelSelect'
-import { AIAgentGrpcApi } from '@/pages/ai-re-act/hooks/grpcApi'
-import { AIChatData } from '../../type/aiChat'
-import { AIDetailsDashIcon } from '../../aiChatWelcome/icon'
-import { Tooltip } from 'antd'
-import useAIGlobalConfig from '@/pages/ai-re-act/hooks/useAIGlobalConfig'
-import ContextTable from './ContextTable/ContextTable'
-import { useI18nNamespaces } from '@/i18n/useI18nNamespaces'
-import { YakitRadioButtons } from '@/components/yakitUI/YakitRadioButtons/YakitRadioButtons'
-import { QuestionCircleOutlined } from '@ant-design/icons'
 
 const AIContextToken: FC<{
   session?: string
@@ -48,479 +14,15 @@ const AIContextToken: FC<{
   /** 详情 icon 对应 YakitButton 的 props */
   buttonProps?: Omit<YakitButtonProp, 'icon' | 'children'>
 }> = ({ session, execute, iconOnly, buttonProps }) => {
-  const { t } = useI18nNamespaces(['aiAgent', 'yakitUi'])
-  const { chatIPCEvents } = useChatIPCDispatcher()
-  const [visible, setVisible] = useState<boolean>(false)
-  const getPerfData = useCallback(() => {
-    const store = chatIPCEvents.fetchChatDataStore?.() ?? aiChatDataStore
-    const data = store?.get(session ?? '')?.aiPerfData ?? null
-    return data
-  }, [session, chatIPCEvents])
-
-  const { renderNumber, aiDataRef } = useRafPolling({
-    getData: getPerfData,
-    interval: 2000,
-    shouldStop: () => !execute,
-    resetDeps: [execute],
-    // 优化：如果是一样的数据结构就不更新
-    shouldUpdate: (prev, next) => {
-      if (!prev) return !!next
-      if (!next) return false
-      return isPerfDataChanged(prev, next)
-    },
-    // 进行数据克隆，确保引用变化
-    clone: (data) => cloneDeep(data),
-  })
-  // 上下文压力集合
-  const currentPressuresEcharts: ContextPressureEchartsProps['dataEcharts'] = useCreation(() => {
-    return getPressuresData(aiDataRef?.pressure, 100)
-  }, [renderNumber, aiDataRef?.pressure])
-
-  // 上下文压力预设值
-  const pressureThreshold = useCreation(() => {
-    return getThreshold(aiDataRef?.pressure)
-  }, [renderNumber, aiDataRef?.pressure])
-
-  // 首字符延迟集合
-  const currentCostEcharts: ResponseSpeedEchartsProps['dataEcharts'] = useCreation(() => {
-    return getCostData(aiDataRef?.firstCost, 100)
-  }, [renderNumber, aiDataRef?.firstCost])
-  // AI的Token消耗
-  const token = useCreation(() => {
-    const { consumption } = aiDataRef || {}
-    if (isEmpty(consumption)) return [0, 0, 0]
-    const input = consumption?.input_consumption || 0
-    const output = consumption?.output_consumption || 0
-    const cacheHit = consumption?.cache_hit_token || 0
-    return [formatNumberUnits(input), formatNumberUnits(output), formatNumberUnits(cacheHit)]
-  }, [renderNumber, aiDataRef?.consumption])
-
-  const isShowCost = useCreation(() => {
-    return !!(
-      currentCostEcharts?.data?.intelligent?.length > 0 ||
-      currentCostEcharts?.data?.lightweight?.length > 0 ||
-      currentCostEcharts?.data?.vision?.length > 0
-    )
-  }, [currentCostEcharts.data])
-  const isShowPressure = useCreation(() => {
-    return !!(
-      currentPressuresEcharts?.data?.intelligent?.length > 0 ||
-      currentPressuresEcharts?.data?.lightweight?.length > 0 ||
-      currentPressuresEcharts?.data?.vision?.length > 0
-    )
-  }, [currentPressuresEcharts.data])
-  const maxIntelligentPressure = useCreation(() => {
-    return currentPressuresEcharts.maxValue.intelligent
-  }, [currentPressuresEcharts.maxValue.intelligent])
-  const maxLightweightPressure = useCreation(() => {
-    return currentPressuresEcharts.maxValue.lightweight
-  }, [currentPressuresEcharts.maxValue.lightweight])
-
-  const maxIntelligentCost = useCreation(() => {
-    return currentCostEcharts.maxValue.intelligent
-  }, [currentCostEcharts.maxValue.intelligent])
-  const maxLightweightCost = useCreation(() => {
-    return currentCostEcharts.maxValue.lightweight
-  }, [currentCostEcharts.maxValue.lightweight])
   return (
     <>
-      {!iconOnly && isShowPressure && (
-        <div className={styles['echarts-wrapper']}>
-          <div className={styles['title']}>
-            <span className={styles['text']}>
-              {t('AIContextToken.pressure')}
-              <span className={styles['tip']}>({t('AIContextToken.peak')})</span>
-            </span>
-            {!!maxIntelligentPressure && (
-              <Tooltip title={t('AiAgengt.intelligentModels')}>
-                <span
-                  className={classNames(styles['intelligent'], {
-                    [styles['intelligent-height']]: maxIntelligentPressure > pressureThreshold,
-                  })}
-                >
-                  {formatNumberUnits(maxIntelligentPressure)}
-                </span>
-              </Tooltip>
-            )}
-            {!!maxLightweightPressure && (
-              <Tooltip title={t('AiAgengt.lightweightModels')}>
-                <span
-                  className={classNames(styles['lightweight'], {
-                    [styles['lightweight-height']]: maxLightweightPressure > pressureThreshold,
-                  })}
-                >
-                  {formatNumberUnits(maxLightweightPressure)}
-                </span>
-              </Tooltip>
-            )}
-          </div>
-          <ContextPressureEcharts dataEcharts={currentPressuresEcharts} threshold={pressureThreshold} />
-        </div>
-      )}
-      {!iconOnly && isShowCost && (
-        <div className={styles['echarts-wrapper']}>
-          <div className={styles['title']}>
-            <span className={styles['text']}>
-              {t('AIContextToken.speed')}
-              <span className={styles['tip']}>({t('AIContextToken.peak')})</span>
-            </span>
-            <Tooltip title={t('AiAgengt.intelligentModels')}>
-              {!!maxIntelligentCost && (
-                <span className={classNames(styles['intelligent'])}>{`${maxIntelligentCost}ms`}</span>
-              )}
-            </Tooltip>
-            <Tooltip title={t('AiAgengt.lightweightModels')}>
-              {!!maxLightweightCost && (
-                <span className={classNames(styles['lightweight'])}>{`${maxLightweightCost}ms`}</span>
-              )}
-            </Tooltip>
-          </div>
-          <ResponseSpeedEcharts dataEcharts={currentCostEcharts} />
-        </div>
-      )}
-      {!iconOnly && (
-        <div className={styles['info-token']}>
-          <div className={styles['token']}>Tokens:</div>
-          <div className={classNames(styles['token-tag'], styles['upload-token'])}>
-            <OutlineArrowupIcon />
-            {token[0]}
-          </div>
-          <div className={classNames(styles['token-tag'], styles['download-token'])}>
-            <OutlineArrowdownIcon />
-            {token[1]}
-          </div>
-          <div className={classNames(styles['token-tag'], styles['download-token'])}>{token[2]}</div>
-        </div>
-      )}
-      <YakitPopover
-        content={
-          <AIEchartsDetails
-            overallToken={[token[0], token[1], token[2]]}
-            tierConsumption={aiDataRef?.consumption?.tier_consumption}
-            pressure={aiDataRef?.pressure}
-            firstCost={aiDataRef?.firstCost}
-            onClose={() => setVisible(false)}
-            contextStats={aiDataRef?.contextStats}
-            contextSections={aiDataRef?.contextSections}
-            renderNumber={renderNumber}
-          />
-        }
-        destroyTooltipOnHide={true}
-        trigger="click"
-        overlayClassName={styles['echarts-details-popover']}
-        visible={visible}
-        onVisibleChange={setVisible}
-      >
-        <Tooltip title={t('YakitButton.viewDetail')}>
-          <YakitButton isHover={visible} icon={<OutlinePresentationchartlineIcon />} type="outline2" {...buttonProps} />
-        </Tooltip>
-      </YakitPopover>
-      {!iconOnly && <div className={styles['divider-style']}></div>}
+      {!iconOnly && <ContextPressurePanel session={session} execute={execute} />}
+      {!iconOnly && <ContextCostPanel session={session} execute={execute} />}
+      {!iconOnly && <ContextTokenSummary session={session} execute={execute} />}
+      <ContextDetailPopover session={session} execute={execute} buttonProps={buttonProps} />
+      {!iconOnly && <div className={styles['divider-style']} />}
     </>
   )
 }
+
 export default memo(AIContextToken)
-
-interface CurrentModel {
-  /**高质模型 */
-  intelligentModels?: AIModelConfig
-  /**轻量模型 */
-  lightweightModels?: AIModelConfig
-}
-interface AIEchartsDetailsProps {
-  overallToken: [number | string, number | string, number | string]
-  /** ref */
-  tierConsumption?: AIAgentGrpcApi.Consumption['tier_consumption']
-  /** ref */
-  pressure?: AIChatData['aiPerfData']['pressure']
-  /** ref */
-  firstCost?: AIChatData['aiPerfData']['firstCost']
-  /** 上下文字节统计 */
-  contextStats?: AIChatData['aiPerfData']['contextStats']
-  // 上下文成分
-  contextSections?: AIChatData['aiPerfData']['contextSections']
-  onClose: () => void
-  renderNumber: number
-}
-const AIEchartsDetails: React.FC<AIEchartsDetailsProps> = memo((props) => {
-  const { overallToken, tierConsumption, pressure, firstCost, contextStats, contextSections, onClose, renderNumber } =
-    props
-  const { t, i18n } = useI18nNamespaces(['aiAgent'])
-  const ref = useRef<HTMLDivElement>(null)
-  const [inViewport = true] = useInViewport(ref)
-
-  const [aiGlobalConfigData, event] = useAIGlobalConfig()
-
-  useEffect(() => {
-    inViewport && event.onRefresh()
-  }, [event, inViewport])
-  const aiGlobalConfig = useCreation(() => aiGlobalConfigData.aiGlobalConfig, [aiGlobalConfigData.aiGlobalConfig])
-
-  const currentModel = useCreation(() => {
-    let data: CurrentModel = {}
-    if (!!aiGlobalConfig?.IntelligentModels?.length) {
-      data.intelligentModels = aiGlobalConfig.IntelligentModels[0]
-    }
-    if (!!aiGlobalConfig?.LightweightModels?.length) {
-      data.lightweightModels = aiGlobalConfig.LightweightModels[0]
-    }
-    return data
-  }, [aiGlobalConfig.IntelligentModels, aiGlobalConfig.LightweightModels])
-
-  const [contextStatsMetric, setContextStatsMetric] = useState<ContextStatsChartMetric>('tokens')
-
-  const intelligentToken = useCreation(() => {
-    if (!tierConsumption?.intelligent) return [0, 0, 0, 0]
-    const input = tierConsumption.intelligent.input_consumption || 0
-    const output = tierConsumption.intelligent.output_consumption || 0
-    const cacheHit = tierConsumption.intelligent.cache_hit_token || 0
-    let percent = 0
-    if (input !== 0 && cacheHit !== 0) {
-      percent = Number(((Number(cacheHit) / Number(input)) * 100).toFixed(2))
-    }
-    return [formatNumberUnits(input), formatNumberUnits(output), formatNumberUnits(cacheHit), percent]
-  }, [renderNumber, tierConsumption?.intelligent])
-
-  const lightweightToken = useCreation(() => {
-    if (!tierConsumption?.lightweight) return [0, 0, 0, 0]
-    const input = tierConsumption.lightweight.input_consumption || 0
-    const output = tierConsumption.lightweight.output_consumption || 0
-    const cacheHit = tierConsumption.lightweight.cache_hit_token || 0
-    let percent = 0
-    if (input !== 0 && cacheHit !== 0) {
-      percent = Number(((Number(cacheHit) / Number(input)) * 100).toFixed(2))
-    }
-    return [formatNumberUnits(input), formatNumberUnits(output), formatNumberUnits(cacheHit), percent]
-  }, [renderNumber, tierConsumption?.lightweight])
-
-  // 上下文压力集合
-  const pressuresEcharts: AIPressureDetailsEchartsProps['dataEcharts'] = useCreation(() => {
-    return getPressuresData(pressure)
-  }, [renderNumber, pressure?.intelligent, pressure?.lightweight])
-  // 首字符延迟集合
-  const costEcharts: AICostDetailsEchartsProps['dataEcharts'] = useCreation(() => {
-    return getCostData(firstCost)
-  }, [renderNumber, firstCost?.intelligent, firstCost?.lightweight])
-  // 上下文压力预设值
-  const threshold = useCreation(() => {
-    return getThreshold(pressure)
-  }, [renderNumber, pressure?.intelligent, pressure?.lightweight])
-  // 上下文字节统计数据
-  const contextStatsData = useCreation(() => {
-    return getContextStatsData(contextStats?.data)
-  }, [renderNumber, contextStats])
-
-  const contextStatsTotalDisplay = useCreation(() => {
-    if (contextStatsMetric === 'tokens') {
-      return formatNumberUnits(contextStats?.prompt_tokens ?? 0)
-    }
-    return formatNumberUnits(contextStats?.prompt_bytes ?? 0)
-  }, [contextStatsMetric, contextStats?.prompt_bytes, contextStats?.prompt_tokens])
-
-  // 上下文成分
-  const contextSectionsData = useCreation(() => {
-    if (!contextSections?.sections)
-      return {
-        summary: new Map<string, string>(),
-        sections: [],
-      }
-    return {
-      summary: contextSections.summary,
-      sections: contextSections.sections,
-    }
-  }, [renderNumber, contextSections?.sections])
-
-  const getEchartsHeard = useMemoizedFn((title: string) => {
-    return (
-      <div className={styles['echarts-heard']}>
-        <div className={styles['echarts-heard-left']}>
-          <div className={styles['title']}>{title}</div>
-          {title === t('AIContextToken.pressure') && (
-            <div className={styles['threshold']}>
-              {t('AIContextToken.limit')}:{formatNumberUnits(threshold)}
-            </div>
-          )}
-        </div>
-        <div className={styles['extra']}>
-          <div className={styles['intelligent']}>
-            <AIDetailsDashIcon className={styles['intelligent-icon']} />
-            <span>{t('AiAgengt.intelligentModels')}</span>
-          </div>
-          <div className={styles['lightweight']}>
-            <AIDetailsDashIcon className={styles['lightweight-icon']} />
-            <span>{t('AiAgengt.lightweightModels')}</span>
-          </div>
-        </div>
-      </div>
-    )
-  })
-  const isShowPressure = useCreation(() => {
-    return !!(
-      pressuresEcharts?.data?.intelligent?.length > 0 ||
-      pressuresEcharts?.data?.lightweight?.length > 0 ||
-      pressuresEcharts?.data?.vision?.length > 0
-    )
-  }, [pressuresEcharts.data])
-  const isShowCost = useCreation(() => {
-    return !!(
-      costEcharts?.data?.intelligent?.length > 0 ||
-      costEcharts?.data?.lightweight?.length > 0 ||
-      costEcharts?.data?.vision?.length > 0
-    )
-  }, [costEcharts.data])
-  return (
-    <div
-      className={styles['echarts-details-wrapper']}
-      ref={ref}
-      style={{ width: i18n.language.startsWith('zh') ? 480 : 550 }}
-    >
-      <div className={styles['echarts-details-heard']}>
-        <div className={styles['echarts-details-title']}>
-          <OutlinePresentationchartlineIcon />
-          <span>{t('AIContextToken.dataDetails')}</span>
-        </div>
-        <YakitButton icon={<OutlineXIcon />} type="text2" onClick={onClose} />
-      </div>
-      <div className={styles['echarts-details-content']}>
-        <div className={styles['token-wrapper']}>
-          <div className={styles['token-heard']}>
-            <span>Tokens:</span>
-            <div className={styles['token-overall-wrapper']}>
-              <div className={styles['token-overall']}>
-                <span>{t('AIContextToken.totalInput')}</span>
-                <div className={classNames(styles['token-tag'], styles['upload-token'])}>
-                  <OutlineArrowupIcon />
-                  {overallToken[0]}
-                </div>
-              </div>
-              <div className={styles['token-overall']}>
-                <span>{t('AIContextToken.totalOutput')}</span>
-                <div className={classNames(styles['token-tag'], styles['download-token'])}>
-                  <OutlineArrowdownIcon />
-                  {overallToken[1]}
-                </div>
-              </div>
-              <div className={styles['token-overall']}>
-                <span>{t('AIContextToken.cache')}</span>
-                <div className={classNames(styles['token-tag'], styles['download-token'])}>{overallToken[2]}</div>
-              </div>
-            </div>
-          </div>
-          <div className={styles['token-content']}>
-            <AITokens
-              modelType={t('AiAgengt.intelligentModels')}
-              aiModel={currentModel?.intelligentModels}
-              token={[intelligentToken[0], intelligentToken[1], intelligentToken[2], intelligentToken[3]]}
-            />
-            <AITokens
-              modelType={t('AiAgengt.lightweightModels')}
-              aiModel={currentModel?.lightweightModels}
-              token={[lightweightToken[0], lightweightToken[1], lightweightToken[2], lightweightToken[3]]}
-            />
-          </div>
-        </div>
-        {isShowPressure && (
-          <div className={styles['pressure-wrapper']}>
-            {getEchartsHeard(t('AIContextToken.pressure'))}
-            <AIPressureDetailsEcharts dataEcharts={pressuresEcharts} threshold={threshold} />
-          </div>
-        )}
-        {isShowCost && (
-          <div className={styles['cost-wrapper']}>
-            {getEchartsHeard(t('AIContextToken.speed'))}
-            <AICostDetailsEcharts dataEcharts={costEcharts} />
-          </div>
-        )}
-        {(!!contextStats?.prompt_bytes || !!contextStats?.prompt_tokens) && (
-          <div className={styles['cost-wrapper']}>
-            <div className={styles['echarts-heard']}>
-              <div className={styles['echarts-heard-left']}>
-                <div className={styles['title']}>上下文统计</div>
-                <YakitRadioButtons
-                  size="small"
-                  value={contextStatsMetric}
-                  buttonStyle="solid"
-                  onChange={(v) => {
-                    setContextStatsMetric(v.target.value as ContextStatsChartMetric)
-                  }}
-                  options={[
-                    { value: 'tokens', label: 'Token' },
-                    { value: 'bytes', label: 'Byte' },
-                  ]}
-                />
-              </div>
-              <div className={styles['total']}>
-                总数 <span>{contextStatsTotalDisplay}</span>
-              </div>
-            </div>
-            <TokenCountEcharts contextStatsData={contextStatsData} metric={contextStatsMetric} />
-          </div>
-        )}
-        {contextSectionsData?.sections.length > 0 && (
-          <div style={{ height: '320px' }}>
-            <ContextTable contextSectionsData={contextSectionsData} roleLabelMap={contextStats?.data?.role_labels} />
-          </div>
-        )}
-      </div>
-    </div>
-  )
-})
-
-interface AITokensProps {
-  modelType: string
-  aiModel?: AIModelConfig
-  token: [number | string, number | string, number | string, number | string]
-}
-const AITokens: React.FC<AITokensProps> = memo((props) => {
-  const { modelType, aiModel, token } = props
-  const { t } = useI18nNamespaces(['aiAgent'])
-  const icon = useCreation(() => {
-    if (!aiModel?.Provider?.Type) return <></>
-    return getIconByAI(aiModel?.Provider?.Type)
-  }, [aiModel?.Provider?.Type])
-  const modelName = useCreation(() => {
-    return aiModel?.ModelName || ''
-  }, [aiModel?.ModelName])
-  return (
-    <div className={styles['ai-tokens']}>
-      <div className={styles['ai-tokens-heard']}>
-        <span className={styles['title']}>{modelType}</span>
-        <div className={styles['model']}>
-          {icon}
-          <div className={styles['model-text']} title={modelName}>
-            {modelName}
-          </div>
-        </div>
-      </div>
-      <div className={styles['ai-tokens-content']}>
-        <div className={styles['ai-tokens-item']}>
-          <div className={styles['token-item']}>
-            {t('AIContextToken.input')}
-            <OutlineArrowupIcon />
-          </div>
-          <div className={classNames(styles['token-tag'], styles['upload-token'])}>{token[0]}</div>
-        </div>
-        <div className={styles['diver']} />
-        <div className={styles['ai-tokens-item']}>
-          <div className={styles['token-item']}>
-            {t('AIContextToken.output')}
-            <OutlineArrowdownIcon />
-          </div>
-          <div className={classNames(styles['token-tag'], styles['download-token'])}>{token[1]}</div>
-        </div>
-        <div className={styles['diver']} />
-        <div className={styles['ai-tokens-item']}>
-          <div className={styles['token-item']}>
-            {t('AIContextToken.cache')}
-            {Number(token[3]) > 0 && (
-              <Tooltip title={`${t('AIContextToken.cacheTooltip')}：${token[3]}%`}>
-                <QuestionCircleOutlined />
-              </Tooltip>
-            )}
-          </div>
-          <div className={classNames(styles['token-tag'], styles['download-token'])}>{token[2]}</div>
-        </div>
-      </div>
-    </div>
-  )
-})
